Update Positions On Layout, Linting Fixes
